+++ resources/js/components/RoomTabHistory.vue
@@ -1,10 +1,25 @@
 <template>
   <div>
 
-    <div class="flex flex-wrap gap-2 justify-content-end flex-row">
-      <!-- Reload list -->
-      <div class="flex justify-content-end">
+    <div class="flex justify-content-between flex-column lg:flex-row align-items-start gap-2 px-2">
+      <div>
+        <!-- Search field, currently not implemented -->
+      </div>
+      <div class="w-full lg:w-auto flex-grow-1 lg:flex-grow-0 flex justify-content-between flex-wrap align-items-start gap-2">
+        <div class="flex gap-2">
+          <InputGroup class="w-auto">
+            <InputGroupAddon>
+              <i class="fa-solid fa-sort"></i>
+            </InputGroupAddon>
+            <Dropdown :disabled="isBusy" v-model="sortField" :options="sortFields" @change="loadData(1)" option-label="name" option-value="value" />
+            <InputGroupAddon class="p-0">
+              <Button :disabled="isBusy" :icon="sortOrder === 1 ? 'fa-solid fa-arrow-up-short-wide' : 'fa-solid fa-arrow-down-wide-short'" @click="toggleSortOrder" severity="secondary" text class="border-noround-left"  />
+            </InputGroupAddon>
+          </InputGroup>
+        </div>
+        <!-- Reload -->
         <Button
+          class="flex-shrink-0"
           v-tooltip="$t('app.reload')"
           severity="secondary"
           :disabled="isBusy"
@@ -15,68 +30,71 @@
     </div>
 
     <!-- List of all meetings -->
-    <DataTable
-      :totalRecords="paginator.getTotalRecords()"
-      :rows="paginator.getRows()"
-      :first="paginator.getFirst()"
-      :value="meetings"
-      dataKey="id"
-      paginator
-      :paginator-template="paginator.getTemplate()"
-      :current-page-report-template="paginator.getCurrentPageReportTemplate()"
-      :loading="isBusy || loadingError"
-      rowHover
-      stripedRows
-      scrollable
-      lazy
-      @page="onPage"
-      class="mt-4 table-auto md:table-fixed"
-    >
-      <template #loading>
+    <OverlayComponent :show="isBusy || loadingError" z-index="1">
+      <template #overlay>
         <LoadingRetryButton :error="loadingError" @reload="loadData()" />
       </template>
-
-      <template #empty>
-        <InlineNote v-if="!isBusy && !loadingError">{{ $t('meetings.no_historical_data') }}</InlineNote>
-      </template>
-
-      <Column field="start" :header="$t('meetings.start')">
-        <template #body="slotProps">
-          {{ $d(new Date(slotProps.data.start),'datetimeShort') }}
-        </template>
-      </Column>
-
-      <Column field="end" :header="$t('meetings.end')">
-        <template #body="slotProps">
-          {{ slotProps.data.end == null ? $t('meetings.now') : $d(new Date(slotProps.data.end),'datetimeShort') }}
-        </template>
-      </Column>
-      <Column
-        :header="$t('app.actions')"
-        class="action-column action-column-2"
+      <DataView
+        :totalRecords="paginator.getTotalRecords()"
+        :rows="paginator.getRows()"
+        :first="paginator.getFirst()"
+        :value="meetings"
+        lazy
+        dataKey="id"
+        paginator
+        :paginator-template="paginator.getTemplate()"
+        :current-page-report-template="paginator.getCurrentPageReportTemplate()"
+        rowHover
+        @page="onPage"
+        class="mt-4"
       >
-        <template #body="slotProps">
-          <div>
-            <RoomTabHistoryStatisticButton
-              v-if="slotProps.data.statistical"
-              :room-id="props.room.id"
-              :meeting-id="slotProps.data.id"
-              :start="slotProps.data.start"
-              :end="slotProps.data.end"
-              :room-name="props.room.name"
-            />
-            <RoomTabHistoryAttendanceButton
-              v-if="slotProps.data.attendance && slotProps.data.end != null"
-              :room-id="props.room.id"
-              :meeting-id="slotProps.data.id"
-              :start="slotProps.data.start"
-              :end="slotProps.data.end"
-              :room-name="props.room.name"
-            />
+        <!-- Show message on empty list -->
+        <template #empty>
+          <div class="px-2">
+            <InlineNote v-if="!isBusy && !loadingError">{{ $t('meetings.no_historical_data') }}</InlineNote>
           </div>
         </template>
-      </Column>
-    </DataTable>
+
+        <template #list="slotProps">
+          <div class="px-2 border-top-1 border-bottom-1 surface-border">
+            <div v-for="(item, index) in slotProps.items" :key="index">
+              <div class="flex flex-column md:flex-row justify-content-between gap-3 py-3" :class="{ 'border-top-1 surface-border': index !== 0 }">
+                <div class="flex flex-column gap-2">
+                  <p class="text-lg font-semibold m-0">{{ $d(new Date(item.start),'datetimeShort') }}</p>
+                  <div class="flex flex-column gap-2 align-items-start">
+                    <div class="flex flex-row gap-2">
+                      <i class="fa-solid fa-hourglass" />
+                      <p class="text-sm m-0" v-tooltip.bottom="$d(new Date(item.start),'datetimeShort')+' - '+(item.end == null ? $t('meetings.now') : $d(new Date(item.end),'datetimeShort'))">
+                        {{ dateDiff.format(new Date(item.start), item.end == null ? new Date() : new Date(item.end)) }}
+                      </p>
+                    </div>
+                  </div>
+                </div>
+                <div class="flex-shrink-0 flex flex-row gap-1 align-items-start justify-content-end" >
+                  <RoomTabHistoryStatisticButton
+                    v-if="item.statistical"
+                    :room-id="props.room.id"
+                    :meeting-id="item.id"
+                    :start="item.start"
+                    :end="item.end"
+                    :room-name="props.room.name"
+                  />
+                  <RoomTabHistoryAttendanceButton
+                    v-if="item.attendance && item.end != null"
+                    :room-id="props.room.id"
+                    :meeting-id="item.id"
+                    :start="item.start"
+                    :end="item.end"
+                    :room-name="props.room.name"
+                  />
+                </div>
+              </div>
+            </div>
+          </div>
+        </template>
+      </DataView>
+    </OverlayComponent>
+
     <div id="retentionPeriodInfo">
       <Divider/>
       <b>{{ $t('meetings.retention_period') }}</b><br>
@@ -92,7 +110,9 @@
 <script setup>
 import { useSettingsStore } from '../stores/settings';
 import { useApi } from '../composables/useApi.js';
-import { onMounted, ref } from 'vue';
+import { computed, onMounted, ref } from 'vue';
+import { useDateDiff } from '../composables/useDateDiff.js';
+import { useI18n } from 'vue-i18n';
 import { usePaginator } from '../composables/usePaginator.js';
 
 const props = defineProps({
@@ -101,11 +121,24 @@ const props = defineProps({
 
 const api = useApi();
 const settingsStore = useSettingsStore();
+const dateDiff = useDateDiff();
+const { t } = useI18n();
 const paginator = usePaginator();
 
 const meetings = ref([]);
 const isBusy = ref(false);
 const loadingError = ref(false);
+const sortField = ref('start');
+const sortOrder = ref(0);
+
+const sortFields = computed(() => [
+  { name: t('meetings.start'), value: 'start' }
+]);
+
+const toggleSortOrder = () => {
+  sortOrder.value = sortOrder.value === 1 ? 0 : 1;
+  loadData(1);
+};
 
 /**
  * Loads the current and previous meetings of a given room
@@ -116,7 +149,9 @@ function loadData (page = null) {
 
   const config = {
     params: {
-      page: page || paginator.getCurrentPage()
+      page: page || paginator.getCurrentPage(),
+      sort_by: sortField.value,
+      sort_direction: sortOrder.value === 1 ? 'asc' : 'desc'
     }
   };